# api/openapi.yaml
openapi: 3.0.3
info:
  title: NovaNAS Storage API
  version: 0.1.0
  description: ZFS-based storage control plane for a single-node NAS appliance.
servers:
  - url: /api/v1
paths:
...
  /pools:
    get:
      operationId: listPools
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Pool' }
    post:
      operationId: createPool
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/PoolCreateSpec' }
      responses:
        '202': { $ref: '#/components/responses/Accepted' }
        '400': { $ref: '#/components/responses/BadRequest' }
  /pools/{name}:
    parameters:
      - in: path
        name: name
        required: true
        schema: { type: string }
    get:
      operationId: getPool
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PoolDetail' }
        '404': { $ref: '#/components/responses/NotFound' }
    delete:
      operationId: destroyPool
      responses:
        '202': { $ref: '#/components/responses/Accepted' }
        '404': { $ref: '#/components/responses/NotFound' }
  /pools/{name}/scrub:
    parameters:
      - in: path
        name: name
        required: true
        schema: { type: string }
      - in: query
        name: action
        schema: { type: string, enum: [start, stop] }
    post:
      operationId: scrubPool
      responses:
        '202': { $ref: '#/components/responses/Accepted' }
  /pools/{name}/trim:
    parameters:
      - in: path
        name: name
        required: true
        schema: { type: string }
      - in: query
        name: action
        schema: { type: string, enum: [start, stop] }
    post:
      operationId: trimPool
      requestBody:
        required: false
        content:
          application/json:
            schema: { $ref: '#/components/schemas/PoolTrimRequest' }
      responses:
        '202': { $ref: '#/components/responses/Accepted' }
        '400': { $ref: '#/components/responses/BadRequest' }
  /pools/{name}/properties:
    parameters:
      - in: path
        name: name
        required: true
        schema: { type: string }
    patch:
      operationId: patchPoolProperties
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/PoolPropertiesPatch' }
      responses:
        '202': { $ref: '#/components/responses/Accepted' }
        '400': { $ref: '#/components/responses/BadRequest' }
  /pools/{name}/checkpoint:
    parameters:
      - in: path
        name: name
        required: true
        schema: { type: string }
    post:
      operationId: checkpointPool
      description: Create a pool-wide checkpoint that can be rolled back to.
      responses:
        '202': { $ref: '#/components/responses/Accepted' }
        '400': { $ref: '#/components/responses/BadRequest' }
  /pools/{name}/discard-checkpoint:
    parameters:
      - in: path
        name: name
        required: true
        schema: { type: string }
    post:
      operationId: discardPoolCheckpoint
      description: Discard a previously-created pool checkpoint.
      responses:
        '202': { $ref: '#/components/responses/Accepted' }
        '400': { $ref: '#/components/responses/BadRequest' }
  /pools/{name}/upgrade:
    parameters:
      - in: path
        name: name
        required: true
        schema: { type: string }
      - in: query
        name: all
        schema: { type: boolean }
        description: When true, upgrade all pools (`zpool upgrade -a`); the path name is ignored.
    post:
      operationId: upgradePool
      description: Upgrade the on-disk format of the named pool, or all pools when `all=true`.
      responses:
        '202': { $ref: '#/components/responses/Accepted' }
        '400': { $ref: '#/components/responses/BadRequest' }
  /pools/{name}/reguid:
    parameters:
      - in: path
        name: name
        required: true
        schema: { type: string }
    post:
      operationId: reguidPool
      description: Generate a new unique identifier for the pool.
      responses:
        '202': { $ref: '#/components/responses/Accepted' }
        '400': { $ref: '#/components/responses/BadRequest' }
  /pools/sync:
    post:
      operationId: syncPools
      description: Force in-flight transaction groups to disk for the named pools, or all pools when names is empty/omitted. Synchronous.
      requestBody:
        required: false
        content:
          application/json:
            schema: { $ref: '#/components/schemas/PoolSyncRequest' }
      responses:
        '200':
          description: Sync completed
        '400': { $ref: '#/components/responses/BadRequest' }
  /pools/{name}/wait:
    parameters:
      - in: path
        name: name
        required: true
        schema: { type: string }
      - in: query
        name: activity
        required: true
        schema: { type: string, enum: [discard, free, initialize, replace, remove, resilver, scrub, trim] }
      - in: query
        name: timeoutSec
        schema: { type: integer, minimum: 0 }
    post:
      operationId: waitPool
      description: Synchronously block until the named pool activity completes or the timeout elapses.
      responses:
        '200':
          description: Activity completed
        '400': { $ref: '#/components/responses/BadRequest' }
        '504':
          description: Wait timed out
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
  /pools/{name}/metadata:
    parameters:
      - in: path
        name: name
        required: true
        schema: { type: string }
    patch:
      operationId: patchPoolMetadata
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/MetadataPatch' }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ResourceMetadata' }
  /datasets:
    get:
      operationId: listDatasets
      parameters:
        - in: query
          name: pool
          schema: { type: string }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Dataset' }
    post:
      operationId: createDataset
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/DatasetCreateSpec' }
      responses:
        '202': { $ref: '#/components/responses/Accepted' }
  /datasets/{fullname}:
    parameters:
      - in: path
        name: fullname
        required: true
        schema: { type: string, description: URL-encoded dataset name }
    get:
      operationId: getDataset
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/DatasetDetail' }
        '404': { $ref: '#/components/responses/NotFound' }
    patch:
      operationId: patchDatasetProps
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [properties]
              properties:
                properties:
                  type: object
                  additionalProperties: { type: string }
      responses:
        '202': { $ref: '#/components/responses/Accepted' }
    delete:
      operationId: destroyDataset
      parameters:
        - in: query
          name: recursive
          schema: { type: boolean }
      responses:
        '202': { $ref: '#/components/responses/Accepted' }
  /datasets/{fullname}/rename:
    parameters:
      - in: path
        name: fullname
        required: true
        schema: { type: string }
    post:
      operationId: renameDataset
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/DatasetRenameRequest' }
      responses:
        '202': { $ref: '#/components/responses/Accepted' }
        '400': { $ref: '#/components/responses/BadRequest' }
  /datasets/{fullname}/clone:
    parameters:
      - in: path
        name: fullname
        required: true
        schema: { type: string, description: URL-encoded snapshot name (must contain '@') }
    post:
      operationId: cloneDataset
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/DatasetCloneRequest' }
      responses:
        '202': { $ref: '#/components/responses/Accepted' }
        '400': { $ref: '#/components/responses/BadRequest' }
  /datasets/{fullname}/promote:
    parameters:
      - in: path
        name: fullname
        required: true
        schema: { type: string }
    post:
      operationId: promoteDataset
      responses:
        '202': { $ref: '#/components/responses/Accepted' }
  /datasets/{fullname}/load-key:
    parameters:
      - in: path
        name: fullname
        required: true
        schema: { type: string }
    post:
      operationId: loadDatasetKey
      requestBody:
        required: false
        content:
          application/json:
            schema: { $ref: '#/components/schemas/LoadKeyRequest' }
      responses:
        '202': { $ref: '#/components/responses/Accepted' }
  /datasets/{fullname}/unload-key:
    parameters:
      - in: path
        name: fullname
        required: true
        schema: { type: string }
    post:
      operationId: unloadDatasetKey
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                recursive: { type: boolean }
      responses:
        '202': { $ref: '#/components/responses/Accepted' }
  /datasets/{fullname}/change-key:
    parameters:
      - in: path
        name: fullname
        required: true
        schema: { type: string }
    post:
      operationId: changeDatasetKey
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ChangeKeyRequest' }
      responses:
        '202': { $ref: '#/components/responses/Accepted' }
        '400': { $ref: '#/components/responses/BadRequest' }
  /datasets/{fullname}/send:
    parameters:
      - in: path
        name: fullname
        required: true
        schema: { type: string, description: URL-encoded snapshot name }
      - in: query
        name: recursive
        schema: { type: boolean }
      - in: query
        name: raw
        schema: { type: boolean }
      - in: query
        name: compressed
        schema: { type: boolean }
      - in: query
        name: largeBlock
        schema: { type: boolean }
      - in: query
        name: embeddedData
        schema: { type: boolean }
      - in: query
        name: from
        schema: { type: string, description: Incremental base snapshot (`-i`); accepts full or @short name }
    post:
      operationId: sendDataset
      description: Streams the result of `zfs send` as application/octet-stream. Synchronous — not dispatched as a job.
      responses:
        '200':
          description: Binary ZFS send stream
          content:
            application/octet-stream:
              schema: { type: string, format: binary }
        '400': { $ref: '#/components/responses/BadRequest' }
  /datasets/{fullname}/receive:
    parameters:
      - in: path
        name: fullname
        required: true
        schema: { type: string, description: URL-encoded target dataset name }
      - in: query
        name: force
        schema: { type: boolean }
      - in: query
        name: resumable
        schema: { type: boolean }
      - in: query
        name: origin
        schema: { type: string, description: '`-o origin=<snap>` for receiving a clone stream' }
    post:
      operationId: receiveDataset
      description: Pipes the request body into `zfs receive`. Synchronous — not dispatched as a job.
      requestBody:
        required: true
        content:
          application/octet-stream:
            schema: { type: string, format: binary }
      responses:
        '200':
          description: Receive completed
        '400': { $ref: '#/components/responses/BadRequest' }
        '500':
          description: Receive failed
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
  /datasets/{fullname}/bookmark:
    parameters:
      - in: path
        name: fullname
        required: true
        schema: { type: string, description: URL-encoded snapshot name (must contain '@') }
    post:
      operationId: bookmarkSnapshot
      description: Create a bookmark for a snapshot (`zfs bookmark`).
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/DatasetBookmarkRequest' }
      responses:
        '202': { $ref: '#/components/responses/Accepted' }
        '400': { $ref: '#/components/responses/BadRequest' }
  /datasets/{fullname}/destroy-bookmark:
    parameters:
      - in: path
        name: fullname
        required: true
        schema: { type: string, description: URL-encoded dataset name (no '@', no '#') }
    post:
      operationId: destroyDatasetBookmark
      description: |
        Destroy a ZFS bookmark. The URL `{fullname}` is the dataset; the
        request body's `bookmark` field is either the short name (e.g.
        `b1`) — which is combined with the URL dataset to form
        `<dataset>#<short>` — or the full bookmark name `<dataset>#<short>`,
        in which case the dataset prefix must match the URL.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/BookmarkDestroyRequest' }
      responses:
        '202': { $ref: '#/components/responses/Accepted' }
        '400': { $ref: '#/components/responses/BadRequest' }
  /datasets/{fullname}/diff:
    parameters:
      - in: path
        name: fullname
        required: true
        schema: { type: string, description: URL-encoded from-snapshot name (must contain '@') }
      - in: query
        name: to
        schema: { type: string, description: Optional second argument; either a snapshot or a dataset name }
    get:
      operationId: diffDataset
      description: Return the result of `zfs diff -H <from> [<to>]`. Synchronous.
      responses:
        '200':
          description: Diff entries
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/DatasetDiffEntry' }
        '400': { $ref: '#/components/responses/BadRequest' }
  /datasets/{fullname}/bookmarks:
    parameters:
      - in: path
        name: fullname
        required: true
        schema: { type: string, description: URL-encoded dataset name }
    get:
      operationId: listDatasetBookmarks
      description: List bookmarks under a dataset (`zfs list -t bookmark -r`). Synchronous.
      responses:
        '200':
          description: Bookmark list
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Bookmark' }
        '400': { $ref: '#/components/responses/BadRequest' }
  /datasets/{fullname}/metadata:
    parameters:
      - in: path
        name: fullname
        required: true
        schema: { type: string }
    patch:
      operationId: patchDatasetMetadata
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/MetadataPatch' }
      responses:
        '200':
          description: OK
  /datasets/{fullname}/rollback:
    parameters:
      - in: path
        name: fullname
        required: true
        schema: { type: string }
    post:
      operationId: rollbackDataset
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [snapshot]
              properties:
                snapshot: { type: string }
      responses:
        '202': { $ref: '#/components/responses/Accepted' }
  /snapshots:
    get:
      operationId: listSnapshots
      parameters:
        - in: query
          name: dataset
          schema: { type: string }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Snapshot' }
    post:
      operationId: createSnapshot
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [dataset, name]
              properties:
                dataset: { type: string }
                name:    { type: string }
                recursive: { type: boolean, default: false }
      responses:
        '202': { $ref: '#/components/responses/Accepted' }
  /snapshots/{fullname}:
    parameters:
      - in: path
        name: fullname
        required: true
        schema: { type: string }
    delete:
      operationId: destroySnapshot
      responses:
        '202': { $ref: '#/components/responses/Accepted' }
  /snapshots/{fullname}/hold:
    parameters:
      - in: path
        name: fullname
        required: true
        schema: { type: string, description: URL-encoded snapshot name }
    post:
      operationId: holdSnapshot
      description: Place a user hold on a snapshot (`zfs hold`).
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/SnapshotHoldRequest' }
      responses:
        '202': { $ref: '#/components/responses/Accepted' }
        '400': { $ref: '#/components/responses/BadRequest' }
  /snapshots/{fullname}/release:
    parameters:
      - in: path
        name: fullname
        required: true
        schema: { type: string, description: URL-encoded snapshot name }
    post:
      operationId: releaseSnapshot
      description: Release a user hold from a snapshot (`zfs release`).
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/SnapshotReleaseRequest' }
      responses:
        '202': { $ref: '#/components/responses/Accepted' }
        '400': { $ref: '#/components/responses/BadRequest' }
  /snapshots/{fullname}/holds:
    parameters:
      - in: path
        name: fullname
        required: true
        schema: { type: string, description: URL-encoded snapshot name }
    get:
      operationId: listSnapshotHolds
      description: List user holds on a snapshot (`zfs holds -H -p`). Synchronous.
      responses:
        '200':
          description: Holds list
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Hold' }
        '400': { $ref: '#/components/responses/BadRequest' }
  /snapshots/{fullname}/metadata:
    parameters:
      - in: path
        name: fullname
        required: true
        schema: { type: string }
    patch:
      operationId: patchSnapshotMetadata
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/MetadataPatch' }
      responses:
        '200':
          description: OK
...
  /samba/users/{username}:
    parameters:
      - in: path
        name: username
        required: true
        schema: { type: string }
    delete:
      operationId: deleteSambaUser
      responses:
        '202': { $ref: '#/components/responses/Accepted' }
...
components:
  responses:
    Accepted:
      description: Async operation accepted; check Location header for job URL.
      headers:
        Location:
          schema: { type: string }
      content:
        application/json:
          schema:
            type: object
            properties:
              jobId: { type: string, format: uuid }
...
  schemas:
...
    Pool:
      type: object
      properties:
        name:             { type: string }
        sizeBytes:        { type: integer }
        allocated:        { type: integer }
        free:             { type: integer }
        health:           { type: string }
        readOnly:         { type: boolean }
        fragmentationPct: { type: integer }
        capacityPct:      { type: integer }
        dedupRatio:       { type: string }
...
    PoolCreateSpec:
      type: object
      required: [name, vdevs]
      properties:
        name: { type: string }
        vdevs:
          type: array
          items: { $ref: '#/components/schemas/VdevSpec' }
        # Log accepts mirror|stripe|disk vdevs only (raidz/draid are rejected
        # by the API — ZFS does not support them for log vdevs).
        log:   { type: array, items: { $ref: '#/components/schemas/VdevSpec' } }
        cache: { type: array, items: { type: string } }
        spare: { type: array, items: { type: string } }
...